[APP] feat: 지도 데이터 로딩

+++ App/app/src/main/java/com/headmetal/headwareintelligence/ApiService.kt
@@ -43,6 +43,6 @@ interface ApiService {
 
     @GET("/map/marker/{no}")
     suspend fun getAccidentData(
-        @Path("no") no: BigInteger
+        @Path("no") no: Int
     ): AccidentResponse
 }
